requirements.txt
llama-index
llama-index-core
llama-index-vector-stores-chroma
# llama-index-embeddings-openai
# llama-index-embeddings-gemini
llama-index-embeddings-ollama
chromadb-client
fastmcp
python-dotenv
gunicorn
PyYAML